feat: Improved conventions and structure
