+++ data/src/commonMain/kotlin/com/srfmolina/krocy/data/db/KrocyDatabase.kt
@@ -5,7 +5,7 @@ import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.RoomDatabaseConstructor
 import com.srfmolina.krocy.data.db.dao.KrocyItemDao
-import com.srfmolina.krocy.data.entity.KrocyItemEntity
+import com.srfmolina.krocy.data.entity.example.KrocyItemEntity
 
 @Database(entities = [KrocyItemEntity::class], version = 1)
 @ConstructedBy(KrocyDatabaseConstructor::class)
