--- domain/src/commonMain/kotlin/com/srfmolina/krocy/domain/model/KrocyItem.kt
@@ -1,7 +1,0 @@
-package com.srfmolina.krocy.domain.model
-
-data class KrocyItem(
-    val id: Int = 0,
-    val name: String,
-    val createdAt: Long
-)
