+++ domain/src/commonMain/kotlin/com/srfmolina/krocy/domain/model/example/KrocyItem.kt
@@ -1,0 +1,7 @@
+package com.srfmolina.krocy.domain.model.example
+
+data class KrocyItem( // An example
+    val id: Int = 0,
+    val name: String,
+    val createdAt: Long
+)
