+++ domain/src/commonMain/kotlin/com/srfmolina/krocy/domain/repository/KrocyItemRepository.kt
@@ -1,9 +1,9 @@
 package com.srfmolina.krocy.domain.repository
 
-import com.srfmolina.krocy.domain.model.KrocyItem
+import com.srfmolina.krocy.domain.model.example.KrocyItem
 import kotlinx.coroutines.flow.Flow
 
-interface KrocyItemRepository {
+interface KrocyItemRepository { // An example
     fun getAll(): Flow<List<KrocyItem>>
     suspend fun save(item: KrocyItem)
     suspend fun deleteById(id: Int)
