+++ ui/src/commonMain/kotlin/com/srfmolina/krocy/ui/di/UiModule.kt
@@ -8,7 +8,8 @@ import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
 val uiModule = module {
+    singleOf(::ObserveStockUseCase)
+
     viewModelOf(::AppViewModel)
     viewModelOf(::StockViewModel)
-    singleOf(::ObserveStockUseCase)
 }
